rename/fix: package 이름변경(scheme -> plan) 및 타입명 명시/수정
